+++ src/main/scala/beam/agentsim/agents/choice/mode/ModeChoiceDriveIfAvailable.scala
@@ -26,10 +26,5 @@ class ModeChoiceDriveIfAvailable(val beamServices: BeamServices) extends ModeCho
 
   override def utilityOf(alternative: EmbodiedBeamTrip): Double = 0.0
 
-  override def utilityOf(
-    mode: Modes.BeamMode,
-    cost: BigDecimal,
-    time: BigDecimal,
-    numTransfers: Int
-  ): Double = 0.0
+  override def utilityOf(mode: Modes.BeamMode, cost: Double, time: Double, numTransfers: Int): Double = 0.0
 }
