+++ src/test/scala/beam/agentsim/agents/OtherPersonAgentSpec.scala
@@ -98,12 +98,7 @@ class OtherPersonAgentSpec
       Some(alternatives.head)
     override val beamServices: BeamServices = beamSvc
     override def utilityOf(alternative: EmbodiedBeamTrip): Double = 0.0
-    override def utilityOf(
-      mode: BeamMode,
-      cost: BigDecimal,
-      time: BigDecimal,
-      numTransfers: Int
-    ): Double = 0.0
+    override def utilityOf(mode: BeamMode, cost: Double, time: Double, numTransfers: Int): Double = 0.0
   }
 
   // Mock a transit driver (who has to be a child of a mock router)
@@ -152,86 +147,58 @@ class OtherPersonAgentSpec
       vehicles.put(bus.getId, bus)
       vehicles.put(tram.getId, tram)
 
-      val busLeg = EmbodiedBeamLeg(
-        BeamLeg(
-          28800,
-          BeamMode.BUS,
-          600,
-          BeamPath(
-            Vector(),
-            Vector(),
-            Some(TransitStopsInfo(1, Id.createVehicleId("my_bus"), 2)),
-            SpaceTime(new Coord(166321.9, 1568.87), 28800),
-            SpaceTime(new Coord(167138.4, 1117), 29400),
-            1.0
-          )
-        ),
-        Id.createVehicleId("my_bus"),
-        false,
-        None,
-        BigDecimal(0),
-        false
-      )
-      val busLeg2 = EmbodiedBeamLeg(
-        BeamLeg(
-          29400,
-          BeamMode.BUS,
-          600,
-          BeamPath(
-            Vector(),
-            Vector(),
-            Some(TransitStopsInfo(2, Id.createVehicleId("my_bus"), 3)),
-            SpaceTime(new Coord(167138.4, 1117), 29400),
-            SpaceTime(new Coord(180000.4, 1200), 30000),
-            1.0
-          )
-        ),
-        Id.createVehicleId("my_bus"),
-        false,
-        None,
-        BigDecimal(0),
-        false
-      )
-      val tramLeg = EmbodiedBeamLeg(
-        BeamLeg(
-          30000,
-          BeamMode.TRAM,
-          600,
-          BeamPath(
-            Vector(),
-            Vector(),
-            Some(TransitStopsInfo(3, Id.createVehicleId("my_tram"), 4)),
-            SpaceTime(new Coord(180000.4, 1200), 30000),
-            SpaceTime(new Coord(190000.4, 1300), 30600),
-            1.0
-          )
-        ),
-        Id.createVehicleId("my_tram"),
-        false,
-        None,
-        BigDecimal(0),
-        false
-      )
-      val replannedTramLeg = EmbodiedBeamLeg(
-        BeamLeg(
-          35000,
-          BeamMode.TRAM,
-          600,
-          BeamPath(
-            Vector(),
-            Vector(),
-            Some(TransitStopsInfo(3, Id.createVehicleId("my_tram"), 4)),
-            SpaceTime(new Coord(180000.4, 1200), 35000),
-            SpaceTime(new Coord(190000.4, 1300), 35600),
-            1.0
-          )
-        ),
-        Id.createVehicleId("my_tram"),
-        false,
-        None,
-        BigDecimal(0),
-        false
-      )
+      val busLeg = EmbodiedBeamLeg(BeamLeg(
+                28800,
+                BeamMode.BUS,
+                600,
+                BeamPath(
+                  Vector(),
+                  Vector(),
+                  Some(TransitStopsInfo(1, Id.createVehicleId("my_bus"), 2)),
+                  SpaceTime(new Coord(166321.9, 1568.87), 28800),
+                  SpaceTime(new Coord(167138.4, 1117), 29400),
+                  1.0
+                )
+              ), Id.createVehicleId("my_bus"), false, None, 0, false)
+      val busLeg2 = EmbodiedBeamLeg(BeamLeg(
+                29400,
+                BeamMode.BUS,
+                600,
+                BeamPath(
+                  Vector(),
+                  Vector(),
+                  Some(TransitStopsInfo(2, Id.createVehicleId("my_bus"), 3)),
+                  SpaceTime(new Coord(167138.4, 1117), 29400),
+                  SpaceTime(new Coord(180000.4, 1200), 30000),
+                  1.0
+                )
+              ), Id.createVehicleId("my_bus"), false, None, 0, false)
+      val tramLeg = EmbodiedBeamLeg(BeamLeg(
+                30000,
+                BeamMode.TRAM,
+                600,
+                BeamPath(
+                  Vector(),
+                  Vector(),
+                  Some(TransitStopsInfo(3, Id.createVehicleId("my_tram"), 4)),
+                  SpaceTime(new Coord(180000.4, 1200), 30000),
+                  SpaceTime(new Coord(190000.4, 1300), 30600),
+                  1.0
+                )
+              ), Id.createVehicleId("my_tram"), false, None, 0, false)
+      val replannedTramLeg = EmbodiedBeamLeg(BeamLeg(
+                35000,
+                BeamMode.TRAM,
+                600,
+                BeamPath(
+                  Vector(),
+                  Vector(),
+                  Some(TransitStopsInfo(3, Id.createVehicleId("my_tram"), 4)),
+                  SpaceTime(new Coord(180000.4, 1200), 35000),
+                  SpaceTime(new Coord(190000.4, 1300), 35600),
+                  1.0
+                )
+              ), Id.createVehicleId("my_tram"), false, None, 0, false)
 
       val household = householdsFactory.createHousehold(Id.create("dummy", classOf[Household]))
       val population =
@@ -309,49 +276,35 @@ class OtherPersonAgentSpec
         Vector(
           EmbodiedBeamTrip(
             Vector(
-              EmbodiedBeamLeg(
-                BeamLeg(
-                  28800,
-                  BeamMode.WALK,
-                  0,
-                  BeamPath(
-                    Vector(),
-                    Vector(),
-                    None,
-                    SpaceTime(new Coord(166321.9, 1568.87), 28800),
-                    SpaceTime(new Coord(167138.4, 1117), 28800),
-                    1.0
-                  )
-                ),
-                Id.createVehicleId("body-dummyAgent"),
-                true,
-                None,
-                BigDecimal(0),
-                false
-              ),
+              EmbodiedBeamLeg(BeamLeg(
+                                28800,
+                                BeamMode.WALK,
+                                0,
+                                BeamPath(
+                                  Vector(),
+                                  Vector(),
+                                  None,
+                                  SpaceTime(new Coord(166321.9, 1568.87), 28800),
+                                  SpaceTime(new Coord(167138.4, 1117), 28800),
+                                  1.0
+                                )
+                              ), Id.createVehicleId("body-dummyAgent"), true, None, 0, false),
               busLeg,
               busLeg2,
               tramLeg,
-              EmbodiedBeamLeg(
-                BeamLeg(
-                  30600,
-                  BeamMode.WALK,
-                  0,
-                  BeamPath(
-                    Vector(),
-                    Vector(),
-                    None,
-                    SpaceTime(new Coord(167138.4, 1117), 30600),
-                    SpaceTime(new Coord(167138.4, 1117), 30600),
-                    1.0
-                  )
-                ),
-                Id.createVehicleId("body-dummyAgent"),
-                true,
-                None,
-                BigDecimal(0),
-                false
-              )
+              EmbodiedBeamLeg(BeamLeg(
+                                30600,
+                                BeamMode.WALK,
+                                0,
+                                BeamPath(
+                                  Vector(),
+                                  Vector(),
+                                  None,
+                                  SpaceTime(new Coord(167138.4, 1117), 30600),
+                                  SpaceTime(new Coord(167138.4, 1117), 30600),
+                                  1.0
+                                )
+                              ), Id.createVehicleId("body-dummyAgent"), true, None, 0, false)
             )
           )
         ),
@@ -404,26 +357,19 @@ class OtherPersonAgentSpec
           EmbodiedBeamTrip(
             Vector(
               replannedTramLeg,
-              EmbodiedBeamLeg(
-                BeamLeg(
-                  35600,
-                  BeamMode.WALK,
-                  0,
-                  BeamPath(
-                    Vector(),
-                    Vector(),
-                    None,
-                    SpaceTime(new Coord(167138.4, 1117), 35600),
-                    SpaceTime(new Coord(167138.4, 1117), 35600),
-                    1.0
-                  )
-                ),
-                Id.createVehicleId("body-dummyAgent"),
-                true,
-                None,
-                BigDecimal(0),
-                false
-              )
+              EmbodiedBeamLeg(BeamLeg(
+                                35600,
+                                BeamMode.WALK,
+                                0,
+                                BeamPath(
+                                  Vector(),
+                                  Vector(),
+                                  None,
+                                  SpaceTime(new Coord(167138.4, 1117), 35600),
+                                  SpaceTime(new Coord(167138.4, 1117), 35600),
+                                  1.0
+                                )
+                              ), Id.createVehicleId("body-dummyAgent"), true, None, 0, false)
             )
           )
         ),
